# Download chunks as tarball so you can run tests against them.