--- a/pkg/querier/unified_querier.go
+++ b/pkg/querier/unified_querier.go
@@ -40,17 +40,12 @@ type unifiedChunkQuerier struct {
 	csq chunkStoreQuerier
 }
 
-// Select implements storage.Querier.
-func (q *unifiedChunkQuerier) Select(sp *storage.SelectParams, matchers ...*labels.Matcher) (storage.SeriesSet, error) {
-	if sp == nil {
-		return q.metadataQuery(matchers...)
-	}
-
+func (q *unifiedChunkQuerier) Get(ctx context.Context, from, through model.Time, matchers ...*labels.Matcher) ([]chunk.Chunk, error) {
 	css := make(chan []chunk.Chunk, len(q.stores))
 	errs := make(chan error, len(q.stores))
 	for _, store := range q.stores {
 		go func(store ChunkStore) {
-			cs, err := store.Get(q.ctx, model.Time(sp.Start), model.Time(sp.End), matchers...)
+			cs, err := store.Get(ctx, from, through, matchers...)
 			if err != nil {
 				errs <- err
 			} else {
@@ -67,6 +62,19 @@ func (q *unifiedChunkQuerier) Select(sp *storage.SelectParams, matchers ...*labe
 		case cs := <-css:
 			chunks = append(chunks, cs...)
 		}
+	}
+	return chunks, nil
+}
+
+// Select implements storage.Querier.
+func (q *unifiedChunkQuerier) Select(sp *storage.SelectParams, matchers ...*labels.Matcher) (storage.SeriesSet, error) {
+	if sp == nil {
+		return q.metadataQuery(matchers...)
+	}
+
+	chunks, err := q.Get(q.ctx, model.Time(sp.Start), model.Time(sp.End), matchers...)
+	if err != nil {
+		return nil, err
 	}
 
 	return q.csq.partitionChunks(chunks), nil