We're seeing slow PromQL evaluations, so here is some useful code to benchmark them with real data.

Download chunks directly with curl:

```
curl -H X-Scope-OrgID:123 'http://localhost:8080/api/prom/chunks?start=1542800779&end= 1542804379&matcher=node_cpu_seconds_total\{mode="idle"\}' > chunks.tgz
```

Then run the query locally:

```
USERID=123 CHUNKS=$(pwd)/chunks.tgz QUERY='max(100 - (avg by (instance) (irate(node_cpu_seconds_total{mode="idle"}[5m])) * 100))' FROM=1542800779 1542804379=1542671400  STEP=600 go test -cpuprofile=cpu.out -run TestChunkTar ./pkg/querier/
```