@tomwilkie 2018-11-21
Also included some optimisations, see https://github.com/prometheus/tsdb/pull/459