@tomwilkie 2018-11-21
Before these optimisations, my sample query (on ~300MB of chunks) was taking ~10s.  Now it takes 9.7s.  A good days work...