@tomwilkie 2018-11-21
Thats a bit more like it:

```
benchmark                                                                     old ns/op     new ns/op     delta
BenchmarkChunkQueryable/rate(foo[1m])/step_1m0s/DoubleDelta/matrixes/-8       41594898      41124571      -1.13%
BenchmarkChunkQueryable/rate(foo[1m])/step_1m0s/DoubleDelta/iterators/-8      56497547      56868423      +0.66%
BenchmarkChunkQueryable/rate(foo[1m])/step_1m0s/DoubleDelta/batches/-8        36234267      35428971      -2.22%
BenchmarkChunkQueryable/rate(foo[1m])/step_1m0s/Varbit/matrixes/-8            41716092      40411735      -3.13%
BenchmarkChunkQueryable/rate(foo[1m])/step_1m0s/Varbit/iterators/-8           54850733      55323337      +0.86%
BenchmarkChunkQueryable/rate(foo[1m])/step_1m0s/Varbit/batches/-8             36436281      37022962      +1.61%
BenchmarkChunkQueryable/rate(foo[1m])/step_1m0s/Bigchunk/matrixes/-8          57250598      57755876      +0.88%
BenchmarkChunkQueryable/rate(foo[1m])/step_1m0s/Bigchunk/iterators/-8         73049512      72447637      -0.82%
BenchmarkChunkQueryable/rate(foo[1m])/step_1m0s/Bigchunk/batches/-8           50987824      50640485      -0.68%
BenchmarkChunkQueryable/foo/step_1m0s/DoubleDelta/matrixes/-8                 36910377      37038630      +0.35%
BenchmarkChunkQueryable/foo/step_1m0s/DoubleDelta/iterators/-8                50565756      50894520      +0.65%
BenchmarkChunkQueryable/foo/step_1m0s/DoubleDelta/batches/-8                  30535052      30319891      -0.70%
BenchmarkChunkQueryable/foo/step_1m0s/Varbit/matrixes/-8                      36003481      36236088      +0.65%
BenchmarkChunkQueryable/foo/step_1m0s/Varbit/iterators/-8                     48744293      49261676      +1.06%
BenchmarkChunkQueryable/foo/step_1m0s/Varbit/batches/-8                       31397407      31519277      +0.39%
BenchmarkChunkQueryable/foo/step_1m0s/Bigchunk/matrixes/-8                    53005869      52074529      -1.76%
BenchmarkChunkQueryable/foo/step_1m0s/Bigchunk/iterators/-8                   67387683      66719185      -0.99%
BenchmarkChunkQueryable/foo/step_1m0s/Bigchunk/batches/-8                     46041862      45218609      -1.79%
BenchmarkChunkQueryable/rate(foo[1m])/step_10m0s/DoubleDelta/matrixes/-8      22923196      23065282      +0.62%
BenchmarkChunkQueryable/rate(foo[1m])/step_10m0s/DoubleDelta/iterators/-8     8451173       8376139       -0.89%
BenchmarkChunkQueryable/rate(foo[1m])/step_10m0s/DoubleDelta/batches/-8       13066263      13169897      +0.79%
BenchmarkChunkQueryable/rate(foo[1m])/step_10m0s/Varbit/matrixes/-8           21672612      21790516      +0.54%
BenchmarkChunkQueryable/rate(foo[1m])/step_10m0s/Varbit/iterators/-8          11351051      10990156      -3.18%
BenchmarkChunkQueryable/rate(foo[1m])/step_10m0s/Varbit/batches/-8            15028742      15506251      +3.18%
BenchmarkChunkQueryable/rate(foo[1m])/step_10m0s/Bigchunk/matrixes/-8         39408510      37346869      -5.23%
BenchmarkChunkQueryable/rate(foo[1m])/step_10m0s/Bigchunk/iterators/-8        45183221      24823669      -45.06%
BenchmarkChunkQueryable/rate(foo[1m])/step_10m0s/Bigchunk/batches/-8          49639435      29753409      -40.06%
BenchmarkChunkQueryable/foo/step_10m0s/DoubleDelta/matrixes/-8                23778955      24018240      +1.01%
BenchmarkChunkQueryable/foo/step_10m0s/DoubleDelta/iterators/-8               22369754      22004092      -1.63%
BenchmarkChunkQueryable/foo/step_10m0s/DoubleDelta/batches/-8                 24647887      24438749      -0.85%
BenchmarkChunkQueryable/foo/step_10m0s/Varbit/matrixes/-8                     22717918      22611165      -0.47%
BenchmarkChunkQueryable/foo/step_10m0s/Varbit/iterators/-8                    22591078      22473433      -0.52%
BenchmarkChunkQueryable/foo/step_10m0s/Varbit/batches/-8                      23630595      24405037      +3.28%
BenchmarkChunkQueryable/foo/step_10m0s/Bigchunk/matrixes/-8                   39058705      38229155      -2.12%
BenchmarkChunkQueryable/foo/step_10m0s/Bigchunk/iterators/-8                  57114985      38185169      -33.14%
BenchmarkChunkQueryable/foo/step_10m0s/Bigchunk/batches/-8                    66755223      49085947      -26.47%
```

The pathalogical query (in the example above, on 300MB of chunks), has gone from 10s to 0.4s.