@tomwilkie 2018-11-22
Running this in prod now, gave a really nice speed up >2x.